[CI] Split access cohort3 tests into 2 groups
