+++ integration/Makefile
@@ -23,7 +23,7 @@ test:
 	CGO_CFLAGS=$(CRYPTO_FLAG) go test $(if $(VERBOSE),-v,) -coverprofile=$(COVER_PROFILE) $(RACE_FLAG) $(if $(JSON_OUTPUT),-json,) $(if $(NUM_RUNS),-count $(NUM_RUNS),) $(GO_TEST_PACKAGES)
 
 .PHONY: access-tests
-access-tests: access-cohort1-tests access-cohort2-tests access-cohort3-tests
+access-tests: access-cohort1-tests access-cohort2-tests access-cohort3-tests access-cohort4-tests
 
 .PHONY: access-cohort1-tests
 access-cohort1-tests:
@@ -36,6 +36,10 @@ access-cohort2-tests:
 .PHONY: access-cohort3-tests
 access-cohort3-tests:
 	CGO_CFLAGS=$(CRYPTO_FLAG) go test -failfast $(if $(VERBOSE),-v,) $(RACE_FLAG) $(if $(JSON_OUTPUT),-json,) $(if $(NUM_RUNS),-count $(NUM_RUNS),) ./tests/access/cohort3/...
+
+.PHONY: access-cohort4-tests
+access-cohort4-tests:
+	CGO_CFLAGS=$(CRYPTO_FLAG) go test -failfast $(if $(VERBOSE),-v,) $(RACE_FLAG) $(if $(JSON_OUTPUT),-json,) $(if $(NUM_RUNS),-count $(NUM_RUNS),) ./tests/access/cohort4/...
 
 .PHONY: collection-tests
 collection-tests:
