+++ integration/tests/access/cohort4/access_circuit_breaker_test.go
@@ -1,4 +1,4 @@
-package cohort3
+package cohort4
 
 import (
 	"context"
@@ -169,7 +169,7 @@ func (s *AccessCircuitBreakerSuite) TestCircuitBreaker() {
 	// Try to send the transaction for the second time. It should wait less than a second because the circuit breaker
 	// is configured to break after the first failure
 	err = accessClient.SendTransaction(s.ctx, signedTx)
-	//Here we catch the codes.Unknown error, as this is the one that comes from the Circuit Breaker when the state is Open.
+	// Here we catch the codes.Unknown error, as this is the one that comes from the Circuit Breaker when the state is Open.
 	assert.Equal(s.T(), codes.Unknown, status.Code(err))
 
 	// Reconnect the collection node
