+++ integration/tests/access/cohort4/grpc_streaming_blocks_test.go
@@ -1,4 +1,4 @@
-package cohort3
+package cohort4
 
 import (
 	"context"
