+++ integration/tests/access/cohort4/rest_state_stream_test.go
@@ -1,4 +1,4 @@
-package cohort3
+package cohort4
 
 import (
 	"context"
